fix: cast log_level to i32 for Windows compatibility
On Windows, bindgen generates C enums as i32 (MSVC uses signed int),
but log_level was passed as u32, causing a type mismatch in
llama_params_fit(). Cast to i32 since log level values are small
non-negative integers that always fit.

+++ llama-cpp-2/src/model/params.rs
@@ -349,7 +349,7 @@ impl LlamaModelParams {
                 self.buft_overrides.as_mut_ptr(),
                 margins.as_mut_ptr(),
                 n_ctx_min,
-                log_level,
+                log_level as i32,
             )
         };
